requirements.txt
python-dotenv==1.0.0
Flask>=2.0.0
Flask-CORS>=4.0.0
psycopg2-binary>=2.9.0
